requirements.txt
discord.py==2.4.0
alpaca-py==0.24.0
torch==2.3.1
scikit-learn==1.5.0